requirements.txt
fitlog==0.9.13
numpy==1.18.5
pandas==1.0.5
PyYAML==6.0
scikit_learn==1.1.2
torch==1.6.0
transformers==4.20.1